formula save includes window
- move saveFormulaText to SpreadsheetViewportWidget, which tracks the current window.

- Closes https://github.com/mP1/walkingkooka-spreadsheet-react/issues/1599
- save formula not including window = viewport parameter, currently window is absent
